handle application data folder is read only (#1661)
* fixed FileNotFoundError when directory isn't writable (#1640)

 - when using docker, if `user_data_dir()` isn't writable directory,
   `default_data_dir()` use `system temp directory` + `virtualenv`.
   for example, tempdir is `/tmp`, it use `/tmp/virtualenv`

* start making the app-data more explicit and robust


* fix Windows

* fix docs

+++ src/virtualenv/discovery/cached_py_info.py
@@ -14,7 +14,6 @@ import sys
 from collections import OrderedDict
 from hashlib import sha256
 
-from virtualenv.dirs import default_data_dir
 from virtualenv.discovery.py_info import PythonInfo
 from virtualenv.info import PY2, PY3
 from virtualenv.util.path import Path
@@ -25,12 +24,12 @@ from virtualenv.version import __version__
 
 _CACHE = OrderedDict()
 _CACHE[Path(sys.executable)] = PythonInfo()
-_FS_PATH = None
 
 
-def from_exe(cls, exe, raise_on_error=True, ignore_cache=False):
+def from_exe(cls, app_data, exe, raise_on_error=True, ignore_cache=False):
     """"""
-    result = _get_from_cache(cls, exe, ignore_cache=ignore_cache)
+    py_info_cache = _get_py_info_cache(app_data)
+    result = _get_from_cache(cls, py_info_cache, app_data, exe, ignore_cache=ignore_cache)
     if isinstance(result, Exception):
         if raise_on_error:
             raise result
@@ -40,14 +39,21 @@ def from_exe(cls, exe, raise_on_error=True, ignore_cache=False):
     return result
 
 
-def _get_from_cache(cls, exe, ignore_cache=True):
+def _get_py_info_cache(app_data):
+    return None if app_data is None else app_data / "py_info" / __version__
+
+
+def _get_from_cache(cls, py_info_cache, app_data, exe, ignore_cache=True):
     # note here we cannot resolve symlinks, as the symlink may trigger different prefix information if there's a
     # pyenv.cfg somewhere alongside on python3.4+
     exe_path = Path(exe)
     if not ignore_cache and exe_path in _CACHE:  # check in the in-memory cache
         result = _CACHE[exe_path]
+    elif py_info_cache is None:  # cache disabled
+        failure, py_info = _run_subprocess(cls, exe, app_data)
+        result = py_info if failure is None else failure
     else:  # then check the persisted cache
-        py_info = _get_via_file_cache(cls, exe_path, exe)
+        py_info = _get_via_file_cache(cls, py_info_cache, app_data, exe_path, exe)
         result = _CACHE[exe_path] = py_info
     # independent if it was from the file or in-memory cache fix the original executable location
     if isinstance(result, PythonInfo):
@@ -55,15 +61,13 @@ def _get_from_cache(cls, exe, ignore_cache=True):
     return result
 
 
-def _get_via_file_cache(cls, resolved_path, exe):
+def _get_via_file_cache(cls, py_info_cache, app_data, resolved_path, exe):
     key = sha256(str(resolved_path).encode("utf-8") if PY3 else str(resolved_path)).hexdigest()
     py_info = None
     resolved_path_text = ensure_text(str(resolved_path))
     resolved_path_modified_timestamp = resolved_path.stat().st_mtime
-    fs_path = _get_fs_path()
-    data_file = fs_path / "{}.json".format(key)
-
-    with fs_path.lock_for_key(key):
+    data_file = py_info_cache / "{}.json".format(key)
+    with py_info_cache.lock_for_key(key):
         data_file_path = data_file.path
         if data_file_path.exists():  # if exists and matches load
             try:
@@ -72,12 +76,12 @@ def _get_via_file_cache(cls, resolved_path, exe):
                     logging.debug("get PythonInfo from %s for %s", data_file_path, exe)
                     py_info = cls._from_dict({k: v for k, v in data["content"].items()})
                 else:
-                    raise ValueError("force cleanup as stale")
+                    raise ValueError("force close as stale")
             except (KeyError, ValueError, OSError):
                 logging.debug("remove PythonInfo %s for %s", data_file_path, exe)
-                data_file_path.unlink()  # cleanup out of date files
+                data_file_path.unlink()  # close out of date files
         if py_info is None:  # if not loaded run and save
-            failure, py_info = _run_subprocess(cls, exe)
+            failure, py_info = _run_subprocess(cls, exe, app_data)
             if failure is None:
                 file_cache_content = {
                     "st_mtime": resolved_path_modified_timestamp,
@@ -91,22 +95,13 @@ def _get_via_file_cache(cls, resolved_path, exe):
     return py_info
 
 
-def _get_fs_path():
-    global _FS_PATH
-    if _FS_PATH is None:
-        _FS_PATH = default_data_dir() / "py-info" / __version__
-    return _FS_PATH
-
-
-def _run_subprocess(cls, exe):
+def _run_subprocess(cls, exe, app_data):
     resolved_path = Path(os.path.abspath(__file__)).parent / "py_info.py"
-    with ensure_file_on_disk(resolved_path) as resolved_path:
-
-        cmd = [exe, "-s", str(resolved_path)]
+    with ensure_file_on_disk(resolved_path, app_data) as resolved_path:
+        cmd = [exe, str(resolved_path)]
         # prevent sys.prefix from leaking into the child process - see https://bugs.python.org/issue22490
         env = os.environ.copy()
         env.pop("__PYVENV_LAUNCHER__", None)
-
         logging.debug("get interpreter info via cmd: %s", LogCmd(cmd))
         try:
             process = Popen(
@@ -156,14 +151,15 @@ class LogCmd(object):
         return raw
 
 
-def clear():
-    fs_path = _get_fs_path()
-    with fs_path:
-        for filename in fs_path.path.iterdir():
-            if filename.suffix == ".json":
-                with fs_path.lock_for_key(filename.stem):
-                    if filename.exists():
-                        filename.unlink()
+def clear(app_data):
+    py_info_cache = _get_py_info_cache(app_data)
+    if py_info_cache is not None:
+        with py_info_cache:
+            for filename in py_info_cache.path.iterdir():
+                if filename.suffix == ".json":
+                    with py_info_cache.lock_for_key(filename.stem):
+                        if filename.exists():
+                            filename.unlink()
     _CACHE.clear()
